+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/DriverJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.DriverEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DriverJpaRepository extends JpaRepository<DriverEntity,Long> {
+public interface DriverJpaRepository extends JpaRepository<DriverEntity,Integer> {
 
 }
